pyproject.toml
[project]
name = "slack-search"
version = "0.1.0"
description = "Slackベクトル検索ボット"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"slack-bolt>=1.18.0",
	"slack-sdk>=3.26.2",
	"openai>=1.40.0",
	"faiss-cpu>=1.8.0",
	"numpy>=2.3.3",
]

[project.scripts]
slack-vector-bot = "slack_vector_bot.main:run"
